// src/API.ts
export type User = {
  __typename: "User",
  id: string,
  username?: string | null,
  about?: string | null,
  posts?: ModelPostConnection | null,
  imageUrl?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items?:  Array<Post | null > | null,
  nextToken?: string | null,
};

export type Post = {
  __typename: "Post",
  id: string,
  username?: string | null,
  imageUrl?: string | null,
  caption?: string | null,
  user?: User | null,
  likes?: Array< string | null > | null,
  comments?: ModelCommentConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items?:  Array<Comment | null > | null,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  content?: string | null,
  username?: string | null,
  post: Post,
  createdAt: string,
  updatedAt: string,
};
// ...
export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    id: string,
    username?: string | null,
    about?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      items?:  Array< {
        __typename: "Post",
        id: string,
        username?: string | null,
        imageUrl?: string | null,
        caption?: string | null,
        likes?: Array< string | null > | null,
        createdAt: string,
        updatedAt: string,
      } | null > | null,
      nextToken?: string | null,
    } | null,
    imageUrl?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    id: string,
    username?: string | null,
    about?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      items?:  Array< {
        __typename: "Post",
        id: string,
        username?: string | null,
        imageUrl?: string | null,
        caption?: string | null,
        likes?: Array< string | null > | null,
        createdAt: string,
        updatedAt: string,
      } | null > | null,
      nextToken?: string | null,
    } | null,
    imageUrl?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type ListPostsQuery = {
  listPosts?:  {
    __typename: "ModelPostConnection",
    items?:  Array< {
      __typename: "Post",
      id: string,
      username?: string | null,
      imageUrl?: string | null,
      caption?: string | null,
      user?:  {
        __typename: "User",
        id: string,
        username?: string | null,
        about?: string | null,
        imageUrl?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      likes?: Array< string | null > | null,
      comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null > | null,
    nextToken?: string | null,
  } | null,
};
// ...
export type OnUpdateUserSubscription = {
  onUpdateUser?:  {
    __typename: "User",
    id: string,
    username?: string | null,
    about?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      items?:  Array< {
        __typename: "Post",
        id: string,
        username?: string | null,
        imageUrl?: string | null,
        caption?: string | null,
        likes?: Array< string | null > | null,
        createdAt: string,
        updatedAt: string,
      } | null > | null,
      nextToken?: string | null,
    } | null,
    imageUrl?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type OnCreatePostSubscription = {
  onCreatePost?:  {
    __typename: "Post",
    id: string,
    username?: string | null,
    imageUrl?: string | null,
    caption?: string | null,
    user?:  {
      __typename: "User",
      id: string,
      username?: string | null,
      about?: string | null,
      posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
      } | null,
      imageUrl?: string | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    likes?: Array< string | null > | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      items?:  Array< {
        __typename: "Comment",
        id: string,
        content?: string | null,
        username?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null > | null,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type OnUpdatePostSubscription = {
  onUpdatePost?:  {
    __typename: "Post",
    id: string,
    username?: string | null,
    imageUrl?: string | null,
    caption?: string | null,
    user?:  {
      __typename: "User",
      id: string,
      username?: string | null,
      about?: string | null,
      posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
      } | null,
      imageUrl?: string | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    likes?: Array< string | null > | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      items?:  Array< {
        __typename: "Comment",
        id: string,
        content?: string | null,
        username?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null > | null,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type OnCreateCommentSubscription = {
  onCreateComment?:  {
    __typename: "Comment",
    id: string,
    content?: string | null,
    username?: string | null,
    post:  {
      __typename: "Post",
      id: string,
      username?: string | null,
      imageUrl?: string | null,
      caption?: string | null,
      user?:  {
        __typename: "User",
        id: string,
        username?: string | null,
        about?: string | null,
        imageUrl?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      likes?: Array< string | null > | null,
      comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    },
    createdAt: string,
    updatedAt: string,
  } | null,
};